Experiment/src/rag_cli/requirements.txt
typer[all]
langchain
openai
psycopg2-binary
pgvector
apache-age
python-dotenv
tiktoken
